+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 ### Added
 
 ### Changed
+- Improve consistency between Marketplace provider and Provider Component provider form and data model - improvement of User experience (@danielkryska)
 
 ### Deprecated